Refactor Bars crate: rename package to lowercase, update README, and enhance tests
- Changed package name from "Bars" to "bars" in Cargo.toml and Cargo.lock.
- Updated .gitignore to exclude .codex files.
- Modified README.md for clarity and consistency in descriptions of bar types.
- Adjusted Tick struct to use f64 for volume instead of u64 for better precision.
- Refactored BarAggregator and ImbalanceBarAggregator to handle volume as f64.
- Enhanced integration tests to validate bar closing behavior and timestamp alignment.

+++ src/types.rs
@@ -23,13 +23,13 @@ impl Direction {
 pub struct Tick {
     pub timestamp: DateTime<Utc>,
     pub price: f64,
-    pub volume: u64,
+    pub volume: f64,
     #[serde(skip)]
     pub direction: Option<Direction>,
 }
 
 impl Tick {
-    pub fn new(timestamp: DateTime<Utc>, price: f64, volume: u64) -> Self {
+    pub fn new(timestamp: DateTime<Utc>, price: f64, volume: f64) -> Self {
         Self { 
             timestamp, 
             price, 
@@ -50,12 +50,12 @@ pub struct Bar {
     pub high: f64,
     pub low: f64,
     pub close: f64,
-    pub volume: u64,
+    pub volume: f64,
     pub timestamp: DateTime<Utc>,
 }
 
 impl Bar {
-    pub fn new(open: f64, high: f64, low: f64, close: f64, volume: u64, timestamp: DateTime<Utc>) -> Self {
+    pub fn new(open: f64, high: f64, low: f64, close: f64, volume: f64, timestamp: DateTime<Utc>) -> Self {
         Self { open, high, low, close, volume, timestamp }
     }
 }
